Removed RVM dep., updated to Hiera 5 / Puppet 4
Previously this module relied on RVM for the ruby used by the webhook
receiver. Now it uses the one bundled with the Puppet agent.

Changes were also made to utilize Hiera 5 and the Puppet 4 language.
This module now works on Puppet 4 and Puppet 5.

Full testing has also been implemented in preparation for a 1.0.0
release.